Extending the "Join Task" Button
For some projects where only one user is needed to perform a task, the
wording "Claim" can be more intuitive than "Join Task". This change
checks to see whether the number of accepted people on a task is equal
to zero, while the number of needed people is equal to one. If this is
the case, then the text on the Join Task button will be "Claim".

For all other cases, the wording on the task will be "Join Task"