Adding new sections to the Task page
This update will add two new sections to the "Task" page:

1) "Unclaimed Tasks" - These are all the open tasks that would normally
show in the "More Tasks" section, that have 0 participants signed up for
the task. If you are the leader of a task but there are 0 participants,
the task will not show in this section (it will continue to show in the
"Your Tasks" section)

2)"Closed Tasks" - All of the closed tasks for a project will show in
this section.

==================================================
Change to existing behavior (1 change):

1) Closed tasks will no longer show in the "More Tasks" section. They
will now show in the "Closed Tasks" section